Remove Manifest.toml
Manifest.toml is generated by Pkg, the Julia package manager, based on
a corresponding Project.toml.  It records a fixed state of all
packages used by the project.  As such, it should not be committed for
packages, but only for applications that require a static environment.

<https://pkgdocs.julialang.org/v1/toml-files/#Manifest.toml>

This is similar to Cargo.lock

<https://doc.rust-lang.org/cargo/guide/cargo-toml-vs-cargo-lock.html>