Rely on Unitful/Base promotion mechanisms
Since the package was first written, it seems that the promotion
mechanism was more convoluted.  Now that both Julia and Uniful have
matured more, it seems that the promotion of MixedLength works fine.

Notably, we no longer require all underlying number types in
MixedLength to be the same type.  I will confirm with @mlhetland that
this would be expected behaviour or not.